% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusterWordFrequencies.r
\name{clusterWordFrequencies}
\alias{clusterWordFrequencies}
\title{Function clusterWordFrequencies}
\usage{
clusterWordFrequencies(corpus, cor = F, binary = T, min = 0,
  plot = T)
}
\description{
Function clusterWordFrequencies
}
