% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.decisions.r
\name{compare.decisions}
\alias{compare.decisions}
\title{Function compare.decisions}
\usage{
compare.decisions(c1 = 2^(seq(1, 100) - 1), p1 = 1/2^seq(1, 100),
  c2 = 10, p2 = 1, plot = T)
}
\description{
Function compare.decisions
}
