% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulativeProspectTheory.r
\name{cumulativeProspectTheory}
\alias{cumulativeProspectTheory}
\title{Function cumulativeProspectTheory}
\usage{
cumulativeProspectTheory(c = 2^(seq(1, 100) - 1), p = 1/2^seq(1, 100),
  vFun = NULL, w1Fun = NULL, w2Fun = NULL, plot = T)
}
\description{
Function cumulativeProspectTheory
}
