% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eliminationHeuristic.r
\name{eliminationHeuristic}
\alias{eliminationHeuristic}
\title{Function eliminationHeuristic}
\usage{
eliminationHeuristic(c1, p1, c2, p2)
}
\description{
Function eliminationHeuristic
}
