% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leastLikelyHeuristic.r
\name{leastLikelyHeuristic}
\alias{leastLikelyHeuristic}
\title{Function leastLikelyHeuristic}
\usage{
leastLikelyHeuristic(c1, p1, c2, p2)
}
\description{
Function leastLikelyHeuristic
}
