% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexicographicHeuristic.r
\name{lexicographicHeuristic}
\alias{lexicographicHeuristic}
\title{Function lexicographicHeuristic}
\usage{
lexicographicHeuristic(c1, p1, c2, p2)
}
\description{
Function lexicographicHeuristic
}
