% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prospectTheory.r
\name{prospectTheory}
\alias{prospectTheory}
\title{Function prospectTheory}
\usage{
prospectTheory(c = 2^(seq(1, 100) - 1), p = 1/2^seq(1, 100),
  vFun = NULL, wFun = NULL, plot = T)
}
\description{
Function prospectTheory
}
