% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.decisions.r
\name{compare.decisions}
\alias{compare.decisions}
\title{Function compare.decisions}
\usage{
compare.decisions(m = NULL, maxi = 20, goal = 1, extended = F)
}
\description{
Function compare.decisions
}
