% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cumulativeProspectTheory.r
\name{cumulativeProspectTheory}
\alias{cumulativeProspectTheory}
\title{Function cumulativeProspectTheory}
\usage{
cumulativeProspectTheory(p = c(1/6, 1/6, 1/6, 1/6, 1/6, 1/6), k = c(-5,
  -3, -1, 2, 4, 6), vFun = NULL, w1Fun = NULL, w2Fun = NULL)
}
\description{
Function cumulativeProspectTheory
}
